third_party/blink/web_tests/fast/forms/range/range-slow-drag-to-edge-expected.txt

Test for dragging operations of <input type=range>

Dragging to edge should always set the value to min/max.
PASS input.value is "0"
PASS input.value is "255"


PASS successfullyParsed is true

TEST COMPLETE
